entity 동적 생성 방법 조사 중

+++ src/main/java/com/example/Jinus/repository/CollegeRepository.java
@@ -3,5 +3,5 @@ package com.example.Jinus.repository;
 import com.example.Jinus.entity.CollegeEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CollegeRepository extends JpaRepository<CollegeEntity, Long> {
+public interface CollegeRepository extends JpaRepository<CollegeEntity, Integer> {
 }
